Style changes; added jump to capability
Did some restyling including moving filters to top of page on search index. Added some stuff to allow the page to jump to the results on the details page if one of the "compare" buttons is clicked (either from the search index or on the details page itself). This is also reflected in the share link. The header is no longer sticky at phone size (to maximize vertical space on these small screens), but I can change this back if desired.