Removing sliders
There is a bad interaction between some of our code and Algolia's code, resulting in unexpected behavior from their slider widgets. While we wait for an answer from them about this, we've removed the sliders. I also added an IE polyfill and made it so that there is more feedback when a certain refinement isn't available.