Changing spinner to text
Since the animation will always pause when the UI is frozen due to synchronous js running, I replaced the spinner with text informing the user of what's going on. I also moved the loading message to start inside getFullData so it will apply in other cases, not just when the "Compare Libraries" button is clicked.